requirements.txt
file:///Users/shurrey/local/agents-sdk-main
box-sdk-gen
python-dotenv